# apispecs/src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Web Project API
  description: Api connect to Database
  version: 0.0.1
  license:
    name: TDTU Web Project
servers:
  - url: http://XXXXX/v1
    description: Develop server
  - url: http://XXXXX.com
    description: Product server
paths:
...
  /v1/getAllPropertyType:
    get:
      tags:
        - Property
      summary: Api for get all type of property
      operationId: getAllPropertyType
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: All Type Name Of Property
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllPropertyTypeResponse'
  /v1/searchProperty:
    post:
      tags:
        - Property
      summary: Api for search property
      operationId: searchProperty
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchPropertyRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Search property in data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPropertyResponse'
  /v1/createProperty:
    post:
      tags:
        - Property
      summary: Api for create property
      operationId: createProperty
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyCreateResponse'
  /v1/updateProperty:
    post:
      tags:
        - Property
      summary: Api for update info of Property
      operationId: updateProperty
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update Property
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyUpdateResponse'
  /v1/deleteProperty:
    delete:
      tags:
        - Property
      summary: Api for delete property
      operationId: deleteProperty
      parameters:
        - name: propertyId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status delete user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyDeleteResponse'
  /v1/getProperty:
    get:
      tags:
        - Property
      summary: Api for get Property by PropertyId
      operationId: getProperty
      parameters:
        - name: propertyId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyGetResponse'
  /v1/getAllStoreByUser:
    get:
      tags:
        - Store
      summary: Api for get all Store by UserId
      operationId: getAllStoreByUser
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            example: '123456'
          required: true
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Store List
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreGetAllResponse'
  /v1/updateStore:
    post:
      tags:
        - Store
      summary: Api for update Store
      operationId: updateStore
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUpdateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Status update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreUpdateResponse'
...
components:
...
  schemas:
    WebApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/WebApiErrorDetails'
    WebApiErrorDetails:
      type: object
      properties:
        field:
          type: string
        errorCd:
          type: string
        message:
          type: string
        errorIdList:
          type: array
          items:
            type: integer
            example: 1
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 400
        message:
          type: string
...
    AllPropertyTypeResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            resultsTotalCount:
              description: Number of user with condition
              type: integer
              format: int64
              example: 1000
            typeProperties:
              description: List of name type property
              type: array
              items:
                $ref: '#/components/schemas/PropertyTypeResponse'
    PropertyTypeResponse:
      type: object
      properties:
        nameProperty:
          description: Name type type property
          type: string
          example: '123456'
        totalProperty:
          description: Number of property with type
          type: integer
          format: int64
          example: 1000
    SearchPropertyRequest:
      x-custom-validate:
        date-range:
          - from: 'createDatetimeFrom'
            to: 'createDatetimeTo'
      type: object
      properties:
        userId:
          description: ID of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        userName:
          description: userName of user create property
          type: string
          example: '123456'
          maxLength: 200
        typeId:
          description: Type ID of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        title:
          description: Title of property
          type: string
          example: '123456'
          maxLength: 200
        address:
          description: Address of property
          type: string
          example: '123456'
          maxLength: 200
        createDatetimeFrom:
          description: Date time create property
          type: string
          example: '123456'
          pattern: '^([12]\d{3})\/(0[1-9]|1[0-2])\/(0[1-9]|1\d|2[0-8]|29(?=\/\d\d\/(?!1[01345789]00|2[1235679]00)\d\d(?:[02468][048]|[13579][26]))|30(?!\/02)|31(?=\/0[13578]|\/1[02]))$'
        createDatetimeTo:
          description: Date time create property
          type: string
          example: '123456'
          pattern: '^([12]\d{3})\/(0[1-9]|1[0-2])\/(0[1-9]|1\d|2[0-8]|29(?=\/\d\d\/(?!1[01345789]00|2[1235679]00)\d\d(?:[02468][048]|[13579][26]))|30(?!\/02)|31(?=\/0[13578]|\/1[02]))$'
        amountFrom:
          description: Amount of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,9}$'
        amountTo:
          description: Amount of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,9}$'
        areaFrom:
          description: Area of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,9}$'
        areaTo:
          description: Area of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,9}$'
        offset:
          description: The offset rows before beginning to return the rows
          type: integer
          format: int32
          minimum: 0
          example: 1
        limit:
          description: The number of rows
          type: integer
          format: int32
          minimum: 1
          example: 20
    SearchPropertyResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            resultsTotalCount:
              description: Number of property with condition
              type: integer
              format: int64
              example: 1000
            properties:
              description: Property list
              type: array
              items:
                $ref: '#/components/schemas/PropertyResponse'
    PropertyResponse:
      type: object
      properties:
        propertyId:
          description: ID of property
          type: string
          example: '123456'
        typeId:
          description: ID of type property
          type: string
          example: '123456'
        typeName:
          description: Name of type property
          type: string
          example: 'xxxxx'
        createUserId:
          description: ID of user create property
          type: string
          example: '123456'
        createUserName:
          description: Name of user create property
          type: string
          example: 'xxxxx'
        createDatetime:
          description: Date-time create property
          type: string
          example: '2022-12-20 00:00:00'
        lastupUserId:
          description: ID of user lastup property
          type: string
          example: '123456'
        lastupUserName:
          description: Name of user lastup property
          type: string
          example: 'xxxxx'
        lastupDatetime:
          description: Date-time lastup property
          type: string
          example: '2022-12-20 00:00:00'
        title:
          description: title of property
          type: string
          example: '2022-12-20 00:00:00'
        address:
          description: Address of property
          type: string
          example: 'Ho Chi Minh'
        amount:
          description: Amount of property
          type: string
          example: '1000000'
        area:
          description: Area of property
          type: string
          example: '10'
        description:
          description: Description of property
          type: string
          example: 'xxxxx'
        approveStatus:
          description: Approve Status of property
          type: string
          example: 'APPROVE'
        propertyImg:
          description: List of IMG property
          type: array
          items:
            description: Link of img property
            type: string
            example: 'xxxxxx'
    PropertyCreateRequest:
      type: object
      required:
        - typeId
        - createUserId
        - title
        - address
        - amount
        - area
        - description
        - propertyImg
      properties:
        typeId:
            description: ID of type property
            type: string
            example: '123456'
            pattern: '^[0-9]{1,19}$'
        createUserId:
            description: ID of user create property
            type: string
            example: '123456'
            pattern: '^[0-9]{1,19}$'
        title:
          description: Title of property
          type: string
          example: 'xxxxx'
          maxLength: 200
        address:
          description: Address of property
          type: string
          example: 'xxxxx'
          maxLength: 200
        amount:
          description: Amount of property
          type: string
          example: '100000'
          pattern: '^[0-9]{1,9}$'
          maxLength: 9
        area:
          description: Area of property
          type: string
          example: '100000'
          pattern: '^[0-9]{1,9}$'
          maxLength: 9
        description:
          description: Description of property
          type: string
          example: 'xxxxx'
          maxLength: 40000
        propertyImg:
          description: Image list
          type: array
          minItems: 3
          maxItems: 12
          items:
            description: Link of image
            type: string
            example: 'xxxxx'
            maxLength: 200
    PropertyCreateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the create property
          type: string
          example: 'OK'
    PropertyUpdateRequest:
      type: object
      required:
        - propertyId
        - typeId
        - title
        - address
        - amount
        - area
        - description
        - propertyImg
        - lastupUserId
      properties:
        propertyId:
          description: ID of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        typeId:
          description: ID of user
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        title:
          description: Title of property
          type: string
          example: 'xxxxx'
          maxLength: 200
        address:
          description: Address of property
          type: string
          example: 'xxxxx'
          maxLength: 200
        amount:
          description: Amount of property
          type: string
          example: '100000'
          pattern: '^[0-9]{1,9}$'
          maxLength: 9
        area:
          description: Area of property
          type: string
          example: '100000'
          pattern: '^[0-9]{1,9}$'
          maxLength: 9
        description:
          description: Description of property
          type: string
          example: 'xxxxx'
          maxLength: 40000
        propertyImg:
          description: Image list
          type: array
          minItems: 3
          maxItems: 12
          items:
            description: Link of image
            type: string
            example: 'xxxxx'
            maxLength: 200
        lastupUserId:
          description: ID of user to create
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
    PropertyUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the update property
          type: string
          example: 'OK'
    PropertyDeleteResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the delete property
          type: string
          example: 'OK'
    PropertyGetResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/PropertyResponse'
    StoreGetAllResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            resultsTotalCount:
              description: Number of property with condition
              type: integer
              format: int64
              example: 1000
            stores:
              description: Store list
              type: array
              items:
                description: Store List of Us
                type: string
                example: 'xxxxxx'
    StoreUpdateRequest:
      type: object
      required:
        - propertyId
        - createUserId
      properties:
        propertyId:
          description: ID of property
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
        createUserId:
          description: ID of user to create
          type: string
          example: '123456'
          pattern: '^[0-9]{1,19}$'
    StoreUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the update store
          type: string
          example: 'OK'
...
  responses:
    webApiValidationError:
      description: webApi input value validation failure
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/WebApiErrorResponse'
            properties:
              status:
                example: 400
...
    401:
      description: Restrict access
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 401
              error:
                example: Unauthorized
    403:
      description: Access request denied
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 403
              error:
                example: Forbidden
    404:
      description: Page not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 404
              error:
                example: Not Found
    500:
      description: The server is having problems
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 500
              error:
                example: Internal Server Error
...